// app/api/types.ts
export interface ApiResponse<T = void> {
  success: boolean;
  data?: T;
  error?: string;
}
// ...
export type CaretakerResponse = Omit<
  Caretaker,
  "createdAt" | "updatedAt" | "deletedAt"
> & {
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
};

export interface CaretakerCreate {
  loginId: string;
  name: string;
  type?: string;
  inactive?: boolean;
  securityPin: string;
}

export interface CaretakerUpdate extends Partial<CaretakerCreate> {
  id: string;
}